+++ src/lib/detector.ts
